LPS-82008 fix tests
