--- modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/test/java/com/liferay/portal/search/elasticsearch6/internal/search/engine/adapter/document/DeleteDocumentRequestExecutorTest.java
@@ -1,103 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.search.elasticsearch6.internal.search.engine.adapter.document;
-
-import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchFixture;
-import com.liferay.portal.search.elasticsearch6.internal.connection.TestElasticsearchConnectionManager;
-import com.liferay.portal.search.engine.adapter.document.DeleteDocumentRequest;
-
-import org.elasticsearch.action.delete.DeleteRequest;
-import org.elasticsearch.action.delete.DeleteRequestBuilder;
-import org.elasticsearch.action.support.WriteRequest;
-
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- * @author Dylan Rebelak
- */
-public class DeleteDocumentRequestExecutorTest {
-
-	@Before
-	public void setUp() throws Exception {
-		_elasticsearchFixture = new ElasticsearchFixture(
-			DeleteDocumentRequestExecutorTest.class.getSimpleName());
-
-		_elasticsearchFixture.setUp();
-
-		_elasticsearchConnectionManager =
-			new TestElasticsearchConnectionManager(_elasticsearchFixture);
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		_elasticsearchFixture.tearDown();
-	}
-
-	@Test
-	public void testDocumentRequestTranslationWithNoRefresh() {
-		doTestDocumentRequestTranslationTest(
-			false, WriteRequest.RefreshPolicy.NONE);
-	}
-
-	@Test
-	public void testDocumentRequestTranslationWithRefresh() {
-		doTestDocumentRequestTranslationTest(
-			true, WriteRequest.RefreshPolicy.IMMEDIATE);
-	}
-
-	protected void doTestDocumentRequestTranslationTest(
-		boolean refreshPolicy,
-		WriteRequest.RefreshPolicy expectedRefreshPolicy) {
-
-		String id = "1";
-
-		DeleteDocumentRequest deleteDocumentRequest = new DeleteDocumentRequest(
-			_INDEX_NAME, _MAPPING_NAME, id);
-
-		deleteDocumentRequest.setRefresh(refreshPolicy);
-
-		DeleteDocumentRequestExecutorImpl deleteDocumentRequestExecutorImpl =
-			new DeleteDocumentRequestExecutorImpl() {
-				{
-					elasticsearchConnectionManager =
-						_elasticsearchConnectionManager;
-				}
-			};
-
-		DeleteRequestBuilder deleteRequestBuilder =
-			deleteDocumentRequestExecutorImpl.createDeleteRequestBuilder(
-				deleteDocumentRequest);
-
-		DeleteRequest deleteRequest = deleteRequestBuilder.request();
-
-		Assert.assertEquals(
-			expectedRefreshPolicy, deleteRequest.getRefreshPolicy());
-		Assert.assertEquals(_INDEX_NAME, deleteRequest.index());
-		Assert.assertEquals(_MAPPING_NAME, deleteRequest.type());
-		Assert.assertEquals(id, deleteRequest.id());
-	}
-
-	private static final String _INDEX_NAME = "test_request_index";
-
-	private static final String _MAPPING_NAME = "testMapping";
-
-	private ElasticsearchConnectionManager _elasticsearchConnectionManager;
-	private ElasticsearchFixture _elasticsearchFixture;
-
-}
