+++ modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/test/java/com/liferay/portal/search/elasticsearch6/internal/search/engine/adapter/document/ElasticsearchBulkableDocumentRequestTranslatorTest.java
@@ -1,0 +1,273 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.search.elasticsearch6.internal.search.engine.adapter.document;
+
+import com.liferay.portal.kernel.search.Document;
+import com.liferay.portal.kernel.search.DocumentImpl;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchConnectionManager;
+import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchFixture;
+import com.liferay.portal.search.elasticsearch6.internal.connection.TestElasticsearchConnectionManager;
+import com.liferay.portal.search.elasticsearch6.internal.document.DefaultElasticsearchDocumentFactory;
+import com.liferay.portal.search.elasticsearch6.internal.document.ElasticsearchDocumentFactory;
+import com.liferay.portal.search.engine.adapter.document.DeleteDocumentRequest;
+import com.liferay.portal.search.engine.adapter.document.IndexDocumentRequest;
+import com.liferay.portal.search.engine.adapter.document.UpdateDocumentRequest;
+import com.liferay.portal.search.test.util.indexing.DocumentFixture;
+
+import java.io.IOException;
+
+import org.elasticsearch.action.bulk.BulkAction;
+import org.elasticsearch.action.bulk.BulkRequestBuilder;
+import org.elasticsearch.action.delete.DeleteRequest;
+import org.elasticsearch.action.delete.DeleteRequestBuilder;
+import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.action.index.IndexRequestBuilder;
+import org.elasticsearch.action.support.WriteRequest;
+import org.elasticsearch.action.update.UpdateRequest;
+import org.elasticsearch.action.update.UpdateRequestBuilder;
+import org.elasticsearch.client.Client;
+import org.elasticsearch.common.xcontent.XContentHelper;
+import org.elasticsearch.common.xcontent.XContentType;
+
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+/**
+ * @author Michael C. Han
+ */
+public class ElasticsearchBulkableDocumentRequestTranslatorTest {
+
+	@Before
+	public void setUp() throws Exception {
+		_elasticsearchFixture = new ElasticsearchFixture(
+			ElasticsearchBulkableDocumentRequestTranslatorTest.class.
+				getSimpleName());
+
+		_elasticsearchFixture.setUp();
+
+		_elasticsearchConnectionManager =
+			new TestElasticsearchConnectionManager(_elasticsearchFixture);
+
+		_documentFixture.setUp();
+
+		_elasticsearchBulkableDocumentRequestTranslator =
+			new ElasticsearchBulkableDocumentRequestTranslator() {
+				{
+					elasticsearchConnectionManager =
+						_elasticsearchConnectionManager;
+				}
+			};
+	}
+
+	@After
+	public void tearDown() throws Exception {
+		_documentFixture.tearDown();
+
+		_elasticsearchFixture.tearDown();
+	}
+
+	@Test
+	public void testDeleteDocumentRequestTranslationWithNoRefresh() {
+		doTestDeleteDocumentRequestTranslation(
+			false, WriteRequest.RefreshPolicy.NONE);
+	}
+
+	@Test
+	public void testDeleteDocumentRequestTranslationWithRefresh() {
+		doTestDeleteDocumentRequestTranslation(
+			true, WriteRequest.RefreshPolicy.IMMEDIATE);
+	}
+
+	@Test
+	public void testIndexDocumentRequestTranslationWithNoRefresh()
+		throws IOException {
+
+		doTestIndexDocumentRequestTranslation(
+			false, WriteRequest.RefreshPolicy.NONE);
+	}
+
+	@Test
+	public void testIndexDocumentRequestTranslationWithRefresh()
+		throws IOException {
+
+		doTestIndexDocumentRequestTranslation(
+			true, WriteRequest.RefreshPolicy.IMMEDIATE);
+	}
+
+	@Test
+	public void testUpdateDocumentRequestTranslationWithNoRefresh()
+		throws IOException {
+
+		doTestUpdateDocumentRequestTranslation(
+			false, WriteRequest.RefreshPolicy.NONE);
+	}
+
+	@Test
+	public void testUpdateDocumentRequestTranslationWithRefresh()
+		throws IOException {
+
+		doTestUpdateDocumentRequestTranslation(
+			true, WriteRequest.RefreshPolicy.IMMEDIATE);
+	}
+
+	protected void doTestDeleteDocumentRequestTranslation(
+		boolean refreshPolicy,
+		WriteRequest.RefreshPolicy expectedRefreshPolicy) {
+
+		String id = "1";
+
+		DeleteDocumentRequest deleteDocumentRequest = new DeleteDocumentRequest(
+			_INDEX_NAME, _MAPPING_NAME, id);
+
+		deleteDocumentRequest.setRefresh(refreshPolicy);
+
+		DeleteRequestBuilder deleteRequestBuilder =
+			_elasticsearchBulkableDocumentRequestTranslator.translate(
+				deleteDocumentRequest, null);
+
+		DeleteRequest deleteRequest = deleteRequestBuilder.request();
+
+		Assert.assertEquals(
+			expectedRefreshPolicy, deleteRequest.getRefreshPolicy());
+		Assert.assertEquals(_INDEX_NAME, deleteRequest.index());
+		Assert.assertEquals(_MAPPING_NAME, deleteRequest.type());
+		Assert.assertEquals(id, deleteRequest.id());
+
+		Client client = _elasticsearchConnectionManager.getClient();
+
+		BulkRequestBuilder bulkRequestBuilder =
+			BulkAction.INSTANCE.newRequestBuilder(client);
+
+		_elasticsearchBulkableDocumentRequestTranslator.translate(
+			deleteDocumentRequest, bulkRequestBuilder);
+
+		Assert.assertEquals(1, bulkRequestBuilder.numberOfActions());
+	}
+
+	protected void doTestIndexDocumentRequestTranslation(
+			boolean refreshPolicy,
+			WriteRequest.RefreshPolicy expectedRefreshPolicy)
+		throws IOException {
+
+		String id = "1";
+
+		Document document = new DocumentImpl();
+
+		document.addKeyword(Field.TYPE, _MAPPING_NAME);
+		document.addKeyword(Field.UID, id);
+
+		IndexDocumentRequest indexDocumentRequest = new IndexDocumentRequest(
+			_INDEX_NAME, document);
+
+		indexDocumentRequest.setRefresh(refreshPolicy);
+
+		IndexRequestBuilder indexRequestBuilder =
+			_elasticsearchBulkableDocumentRequestTranslator.translate(
+				indexDocumentRequest, null);
+
+		IndexRequest indexRequest = indexRequestBuilder.request();
+
+		Assert.assertEquals(
+			expectedRefreshPolicy, indexRequest.getRefreshPolicy());
+		Assert.assertEquals(_INDEX_NAME, indexRequest.index());
+		Assert.assertEquals(_MAPPING_NAME, indexRequest.type());
+		Assert.assertEquals(id, indexRequest.id());
+
+		ElasticsearchDocumentFactory elasticsearchDocumentFactory =
+			new DefaultElasticsearchDocumentFactory();
+
+		String source = XContentHelper.convertToJson(
+			indexRequest.source(), false, XContentType.JSON);
+
+		Assert.assertEquals(
+			elasticsearchDocumentFactory.getElasticsearchDocument(document),
+			source);
+
+		Client client = _elasticsearchConnectionManager.getClient();
+
+		BulkRequestBuilder bulkRequestBuilder =
+			BulkAction.INSTANCE.newRequestBuilder(client);
+
+		_elasticsearchBulkableDocumentRequestTranslator.translate(
+			indexDocumentRequest, bulkRequestBuilder);
+
+		Assert.assertEquals(1, bulkRequestBuilder.numberOfActions());
+	}
+
+	protected void doTestUpdateDocumentRequestTranslation(
+			boolean refreshPolicy,
+			WriteRequest.RefreshPolicy expectedRefreshPolicy)
+		throws IOException {
+
+		String id = "1";
+
+		Document document = new DocumentImpl();
+
+		document.addKeyword(Field.TYPE, _MAPPING_NAME);
+		document.addKeyword(Field.UID, id);
+
+		UpdateDocumentRequest updateDocumentRequest = new UpdateDocumentRequest(
+			_INDEX_NAME, id, document);
+
+		updateDocumentRequest.setRefresh(refreshPolicy);
+
+		UpdateRequestBuilder updateRequestBuilder =
+			_elasticsearchBulkableDocumentRequestTranslator.translate(
+				updateDocumentRequest, null);
+
+		UpdateRequest updateRequest = updateRequestBuilder.request();
+
+		Assert.assertEquals(
+			expectedRefreshPolicy, updateRequest.getRefreshPolicy());
+		Assert.assertEquals(_INDEX_NAME, updateRequest.index());
+		Assert.assertEquals(id, updateRequest.id());
+
+		ElasticsearchDocumentFactory elasticsearchDocumentFactory =
+			new DefaultElasticsearchDocumentFactory();
+
+		IndexRequest indexRequest = updateRequest.doc();
+
+		String source = XContentHelper.convertToJson(
+			indexRequest.source(), false, XContentType.JSON);
+
+		Assert.assertEquals(
+			elasticsearchDocumentFactory.getElasticsearchDocument(document),
+			source);
+
+		Client client = _elasticsearchConnectionManager.getClient();
+
+		BulkRequestBuilder bulkRequestBuilder =
+			BulkAction.INSTANCE.newRequestBuilder(client);
+
+		_elasticsearchBulkableDocumentRequestTranslator.translate(
+			updateDocumentRequest, bulkRequestBuilder);
+
+		Assert.assertEquals(1, bulkRequestBuilder.numberOfActions());
+	}
+
+	private static final String _INDEX_NAME = "test_request_index";
+
+	private static final String _MAPPING_NAME = "testMapping";
+
+	private final DocumentFixture _documentFixture = new DocumentFixture();
+	private ElasticsearchBulkableDocumentRequestTranslator
+		_elasticsearchBulkableDocumentRequestTranslator;
+	private ElasticsearchConnectionManager _elasticsearchConnectionManager;
+	private ElasticsearchFixture _elasticsearchFixture;
+
+}
