--- modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/test/java/com/liferay/portal/search/elasticsearch6/internal/search/engine/adapter/document/UpdateDocumentRequestExecutorTest.java
@@ -1,139 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.search.elasticsearch6.internal.search.engine.adapter.document;
-
-import com.liferay.portal.kernel.search.Document;
-import com.liferay.portal.kernel.search.DocumentImpl;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch6.internal.connection.ElasticsearchFixture;
-import com.liferay.portal.search.elasticsearch6.internal.connection.TestElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch6.internal.document.DefaultElasticsearchDocumentFactory;
-import com.liferay.portal.search.elasticsearch6.internal.document.ElasticsearchDocumentFactory;
-import com.liferay.portal.search.engine.adapter.document.UpdateDocumentRequest;
-import com.liferay.portal.search.test.util.indexing.DocumentFixture;
-
-import java.io.IOException;
-
-import org.elasticsearch.action.index.IndexRequest;
-import org.elasticsearch.action.support.WriteRequest;
-import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.action.update.UpdateRequestBuilder;
-import org.elasticsearch.common.xcontent.XContentHelper;
-import org.elasticsearch.common.xcontent.XContentType;
-
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- * @author Dylan Rebelak
- */
-public class UpdateDocumentRequestExecutorTest {
-
-	@Before
-	public void setUp() throws Exception {
-		_elasticsearchFixture = new ElasticsearchFixture(
-			UpdateDocumentRequestExecutorTest.class.getSimpleName());
-
-		_elasticsearchFixture.setUp();
-
-		_elasticsearchConnectionManager =
-			new TestElasticsearchConnectionManager(_elasticsearchFixture);
-
-		_documentFixture = new DocumentFixture();
-
-		_documentFixture.setUp();
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		_elasticsearchFixture.tearDown();
-		_documentFixture.tearDown();
-	}
-
-	@Test
-	public void testDocumentRequestTranslationWithNoRefresh()
-		throws IOException {
-
-		doTestDocumentRequestTranslation(
-			false, WriteRequest.RefreshPolicy.NONE);
-	}
-
-	@Test
-	public void testDocumentRequestTranslationWithRefresh() throws IOException {
-		doTestDocumentRequestTranslation(
-			true, WriteRequest.RefreshPolicy.IMMEDIATE);
-	}
-
-	protected void doTestDocumentRequestTranslation(
-			boolean refreshPolicy,
-			WriteRequest.RefreshPolicy expectedRefreshPolicy)
-		throws IOException {
-
-		String id = "1";
-
-		Document document = new DocumentImpl();
-
-		document.addKeyword(Field.TYPE, _MAPPING_NAME);
-		document.addKeyword(Field.UID, id);
-
-		UpdateDocumentRequest updateDocumentRequest = new UpdateDocumentRequest(
-			_INDEX_NAME, id, document);
-
-		updateDocumentRequest.setRefresh(refreshPolicy);
-
-		UpdateDocumentRequestExecutorImpl updateDocumentRequestExecutorImpl =
-			new UpdateDocumentRequestExecutorImpl() {
-				{
-					elasticsearchConnectionManager =
-						_elasticsearchConnectionManager;
-				}
-			};
-
-		UpdateRequestBuilder updateRequestBuilder =
-			updateDocumentRequestExecutorImpl.createUpdateRequestBuilder(
-				updateDocumentRequest);
-
-		UpdateRequest updateRequest = updateRequestBuilder.request();
-
-		Assert.assertEquals(
-			expectedRefreshPolicy, updateRequest.getRefreshPolicy());
-		Assert.assertEquals(_INDEX_NAME, updateRequest.index());
-		Assert.assertEquals(id, updateRequest.id());
-
-		ElasticsearchDocumentFactory elasticsearchDocumentFactory =
-			new DefaultElasticsearchDocumentFactory();
-
-		IndexRequest indexRequest = updateRequest.doc();
-
-		String source = XContentHelper.convertToJson(
-			indexRequest.source(), false, XContentType.JSON);
-
-		Assert.assertEquals(
-			elasticsearchDocumentFactory.getElasticsearchDocument(document),
-			source);
-	}
-
-	private static final String _INDEX_NAME = "test_request_index";
-
-	private static final String _MAPPING_NAME = "testMapping";
-
-	private DocumentFixture _documentFixture;
-	private ElasticsearchConnectionManager _elasticsearchConnectionManager;
-	private ElasticsearchFixture _elasticsearchFixture;
-
-}
